consolidated the face samples into a single python file face_samples.py
